fix: Tuist recycle dependencies 문제 수정 및 UIView+Extensions File Method 추가
