+++ Core/HPCommon/Project.swift
@@ -12,7 +12,6 @@ let common = Project.makeModule(
   name: "HPCommon",
   products: [.framework(.static)],
   dependencies: [
-    .Project.Core.extensions,
     .Project.UI.common
   ]
 )
