+++ Core/HPCommon/Sources/SignUp/SignUpTermsView.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 import HPCommonUI
-import HPExtensions
 import SnapKit
 import Then
